Move DisplayBagVersionEntry to another file
